+++ src/os_sensor_nodelet.cpp
@@ -15,7 +15,6 @@
 #include <pluginlib/class_list_macros.h>
 #include <std_srvs/Empty.h>
 
-#include <tuple>
 #include <chrono>
 
 #include "ouster_ros/PacketMsg.h"